Updated Coulmn Types

+++ database/migrations/2022_12_29_091022_create_listings_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->string('type');
             $table->string('image')->nullable();
             $table->string('description');
-            $table->integer('price');
+            $table->decimal('price');
             $table->integer('views')->default(0)->nullable();
             $table->timestamps();
         });
